+++ core/src/decl/Kokkos_Declare_HPX.hpp
@@ -18,7 +18,7 @@
 #define KOKKOS_DECLARE_HPX_HPP
 
 #if defined(KOKKOS_ENABLE_HPX)
-#include <Kokkos_HPX.hpp>
+#include <HPX/Kokkos_HPX.hpp>
 #include <HPX/Kokkos_HPX_MDRangePolicy.hpp>
 #endif
